requirements.txt
anyio==4.3.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2024.2
cryptography==42.0.5
cssselect2==0.7.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2
django-allauth==0.58.2
django-appconf==1.0.6
django-crispy-forms==2.1
django-select2==8.1.2
docopt==0.6.2
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
h11==0.14.0
h2==4.1.0
hishel==0.0.24
hpack==4.0.0
html5lib==1.1
httpcore==1.0.4
httpx==0.27.0
hyperframe==6.0.1
idna==3.6
inflect==7.0.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mifiel==0.0.13
mistune==3.0.2
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
numpy==1.26.4
oauthlib==3.2.2
oscrypto==1.3.0
packaging==24.0
pandocfilters==1.5.1
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pure-eval==0.2.2
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pyHanko==0.23.1
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pypdf==4.1.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.4.2
referencing==0.34.0
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
svglib==1.5.1
sympy==1.12
tenacity==8.2.3
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.2
urllib3==1.26.0
uuid==1.30
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.6.0
xhtml2pdf==0.2.15
yarg==0.1.9